# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hero's Path API specification
servers:
  - url: /api
    description: Base API path
...
security:
  - BearerAuth: []
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [me]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/preferences:
    get:
      operationId: getPreferences
      tags: [me]
      summary: Get user preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updatePreferences
      tags: [me]
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePreferencesRequest"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
  /places/discover:
    get:
      operationId: getDiscoverFeed
      tags: [places]
      summary: Get the discover feed (all discovered places, deduplicated)
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, unreviewed, favorited, snoozed]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Discover feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverFeedResponse"
  /places/{googlePlaceId}/action:
    post:
      operationId: placeAction
      tags: [places]
      summary: Perform an action on a place (snooze, dismiss, favorite, add to list)
      parameters:
        - name: googlePlaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceActionRequest"
      responses:
        "200":
          description: Action performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDiscoveredPlace"
...
  /lists/{listId}/places:
    get:
      operationId: getListPlaces
      tags: [lists]
      summary: Get places in a list
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Places in list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlacesResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: string
        googleId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        xp:
          type: integer
        level:
          type: integer
        rank:
          type: string
        streakDays:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - googleId
        - displayName
        - xp
        - level
        - rank
        - streakDays
        - createdAt
    UserPreferences:
      type: object
      properties:
        placeTypes:
          type: array
          items:
            type: string
        minRating:
          type: number
          minimum: 0
          maximum: 5
      required:
        - placeTypes
        - minRating
    UpdatePreferencesRequest:
      type: object
      properties:
        placeTypes:
          type: array
          items:
            type: string
        minRating:
          type: number
          minimum: 0
          maximum: 5
...
    Place:
      type: object
      properties:
        googlePlaceId:
          type: string
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        rating:
          type: number
          nullable: true
        types:
          type: array
          items:
            type: string
        photoReference:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        distanceM:
          type: number
          nullable: true
      required:
        - googlePlaceId
        - name
        - lat
        - lng
        - types
    UserDiscoveredPlace:
      type: object
      properties:
        googlePlaceId:
          type: string
        name:
          type: string
        lat:
          type: number
        lng:
          type: number
        rating:
          type: number
          nullable: true
        types:
          type: array
          items:
            type: string
        photoReference:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        firstDiscoveredAt:
          type: string
          format: date-time
        lastDiscoveredAt:
          type: string
          format: date-time
        discoveryCount:
          type: integer
        isDismissed:
          type: boolean
        isFavorited:
          type: boolean
        isSnoozed:
          type: boolean
        snoozedUntil:
          type: string
          format: date-time
          nullable: true
      required:
        - googlePlaceId
        - name
        - lat
        - lng
        - types
        - firstDiscoveredAt
        - lastDiscoveredAt
        - discoveryCount
        - isDismissed
        - isFavorited
        - isSnoozed
    PlacesResponse:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: "#/components/schemas/Place"
      required:
        - places
    DiscoverFeedResponse:
      type: object
      properties:
        places:
          type: array
          items:
            $ref: "#/components/schemas/UserDiscoveredPlace"
        total:
          type: integer
      required:
        - places
        - total
    PlaceActionRequest:
      type: object
      properties:
        action:
          type: string
          enum: [snooze, dismiss, favorite, unfavorite, add_to_list, remove_from_list]
        listId:
          type: string
          nullable: true
        snoozeDays:
          type: integer
          nullable: true
          default: 7
      required:
        - action
...